@@ -9,13 +9,14 @@
     "@vercel/edge-config": "^1.4.0",
     "bootstrap": "^4.5.0",
     "cloudinary": "^2.9.0",
+    "framer-motion": "^10.18.0",
     "gsap": "^3.14.2",
     "nes.css": "^2.2.1",
     "ogl": "^1.0.11",
     "pixel-retroui": "^2.1.0",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-bootstrap": "^2.2.1",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-router-dom": "^5.3.0",
     "react-scripts": "^5.0.1",
     "react-slick": "^0.28.1",
