@@ -1,5 +1,5 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import App from './components/App';
 
 // import 'pixel-retroui/dist/fonts.css'; // Commented out - path doesn't exist in pixel-retroui package
@@ -20,7 +20,8 @@ import './assets/css/nes-custom.css';
 import './assets/css/scroll-animations.css';
 import './assets/css/techstorm-theme.css';
 
-ReactDOM.render(<App />, document.getElementById('root'));
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(<App />);
 
 // ─── Service Worker Registration (PWA) ───────────────────────────────────────
 if ('serviceWorker' in navigator) {
