@@ -2,18 +2,28 @@ import React, { useState } from 'react';
 import { useHistory } from 'react-router-dom';
 import { submitEventRegistration } from '../../../utils/eventRegistrationAPI';
 import Breadcrumb from '../../Utilities/Breadcrumb/Breadcrumb';
+import Stepper, { Step } from '../../Utilities/Stepper/Stepper';
 import './Registration.css';
 import fifaMobileBanner from '../../../assets/img/event_specific_pictures/games/fifa_mobile.png';
+import qrCodeImage from '../../../assets/img/QrCode_For_Payment.jpg.jpeg';
+
+const YEAR_OPTIONS = ['1st Year', '2nd Year', '3rd Year', '4th Year'];
+const COLLEGE_OPTIONS = [
+  'B. P. Poddar Institute of Management & Technology',
+  'Others'
+];
 
 const FifaMobileRegistration = () => {
   const history = useHistory();
   const [formData, setFormData] = useState({
     fullName: '',
     collegeName: '',
+    collegeOther: '',
     department: '',
     yearOfStudy: '',
     contactNumber: '',
     emailAddress: '',
+    collegeIdProof: null,
     fifaUsername: '',
     teamOvr: '',
     deviceModel: '',
@@ -29,8 +39,15 @@ const FifaMobileRegistration = () => {
   const [submitSuccess, setSubmitSuccess] = useState(false);
 
   const handleInputChange = (e) => {
-    const { name, value, type, checked } = e.target;
-    const nextValue = type === 'checkbox' ? checked : value;
+    const { name, value, type, checked, files } = e.target;
+    let nextValue = value;
+    
+    if (type === 'checkbox') {
+      nextValue = checked;
+    } else if (type === 'file') {
+      nextValue = files && files[0] ? files[0] : null;
+    }
+    
     setFormData((prev) => ({ ...prev, [name]: nextValue }));
 
     if (errors[name]) {
@@ -38,34 +55,38 @@ const FifaMobileRegistration = () => {
     }
   };
 
-  const validateForm = () => {
+  const validateStep1 = () => {
     const nextErrors = {};
-
     if (!formData.fullName.trim()) nextErrors.fullName = 'Full Name is required';
-    if (!formData.collegeName.trim()) nextErrors.collegeName = 'College Name is required';
+    if (!formData.collegeName) nextErrors.collegeName = 'College selection is required';
+    if (formData.collegeName === 'Others' && !formData.collegeOther.trim()) {
+      nextErrors.collegeOther = 'Please specify your college name';
+    }
     if (!formData.department.trim()) nextErrors.department = 'Department / Branch is required';
     if (!formData.yearOfStudy.trim()) nextErrors.yearOfStudy = 'Year of Study is required';
-
     if (!formData.contactNumber.trim()) {
       nextErrors.contactNumber = 'Contact Number is required';
     } else if (!/^\d{10,15}$/.test(formData.contactNumber.replace(/\D/g, ''))) {
       nextErrors.contactNumber = 'Enter a valid contact number';
     }
-
     if (!formData.emailAddress.trim()) {
       nextErrors.emailAddress = 'Email Address is required';
     } else if (!/\S+@\S+\.\S+/.test(formData.emailAddress)) {
       nextErrors.emailAddress = 'Invalid email format';
     }
-
+    if (!formData.collegeIdProof) nextErrors.collegeIdProof = 'College ID / Library Card upload is required';
     if (!formData.fifaUsername.trim()) nextErrors.fifaUsername = 'FIFA Mobile Username is required';
     if (!formData.teamOvr.trim()) nextErrors.teamOvr = 'Team OVR is required';
     if (!formData.deviceModel.trim()) nextErrors.deviceModel = 'Device Model is required';
-    
+    setErrors(nextErrors);
+    return Object.keys(nextErrors).length === 0;
+  };
+
+  const validateStep2 = () => {
+    const nextErrors = {};
     if (!formData.paymentMode) nextErrors.paymentMode = 'Payment mode is required';
-    if (!formData.paymentDate) nextErrors.paymentDate = 'Payment date is required';
-    
     if (formData.paymentMode === 'online') {
+      if (!formData.paymentDate) nextErrors.paymentDate = 'Payment date is required';
       if (!formData.transactionId.trim()) {
         nextErrors.transactionId = 'Transaction ID is required for online payment';
       }
@@ -73,26 +94,32 @@ const FifaMobileRegistration = () => {
         nextErrors.paymentReceipt = 'Payment screenshot is required for online payment';
       }
     }
-    if (formData.paymentMode === 'offline' && !formData.paymentReceipt) {
-      nextErrors.paymentReceipt = 'Offline receipt is required for offline payment';
-    }
-    
-    if (!formData.agreeToRules) nextErrors.agreeToRules = 'You must agree to tournament rules';
-    if (!formData.whatsappConfirmed) {
-      nextErrors.whatsappConfirmed = 'Please confirm after joining the WhatsApp group';
-    }
-
     setErrors(nextErrors);
     return Object.keys(nextErrors).length === 0;
   };
 
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-
-    if (!validateForm()) {
-      return;
+  const validateStep3 = () => {
+    const nextErrors = {};
+    if (!formData.agreeToRules) nextErrors.agreeToRules = 'You must agree to tournament rules';
+    if (!formData.whatsappConfirmed) {
+      nextErrors.whatsappConfirmed = 'Please confirm after joining the WhatsApp group';
     }
+    setErrors(nextErrors);
+    return Object.keys(nextErrors).length === 0;
+  };
 
+  const handleStepValidation = (step) => {
+    if (step === 1) {
+      return validateStep1();
+    } else if (step === 2) {
+      return validateStep2();
+    } else if (step === 3) {
+      return validateStep3();
+    }
+    return true;
+  };
+
+  const handleComplete = async () => {
     setIsSubmitting(true);
 
     try {
@@ -143,295 +170,443 @@ const FifaMobileRegistration = () => {
             </div>
           )}
 
-          <form className="registration-form" onSubmit={handleSubmit}>
-            <div className="form-section">
-              <h2 className="form-section-title">&gt;&gt;&gt; Participant Details</h2>
+          <Stepper
+            initialStep={1}
+            onStepChange={(step) => console.log('Step:', step)}
+            onStepValidation={handleStepValidation}
+            onFinalStepCompleted={handleComplete}
+            backButtonText="← Previous"
+            nextButtonText="Next Step →"
+          >
+            {/* Step 1: Participant Details */}
+            <Step>
+              <div className="form-section">
+                <h2 className="form-section-title">&gt;&gt;&gt; Participant Details</h2>
 
-              <div className="form-group">
-                <label className="form-label required">Full Name</label>
-                <input
-                  type="text"
-                  name="fullName"
-                  value={formData.fullName}
-                  onChange={handleInputChange}
-                  className="retro-input"
-                  placeholder="Full Name"
-                />
-                {errors.fullName && <div className="error-message">{errors.fullName}</div>}
-              </div>
-
-              <div className="form-group">
-                <label className="form-label required">College Name</label>
-                <input
-                  type="text"
-                  name="collegeName"
-                  value={formData.collegeName}
-                  onChange={handleInputChange}
-                  className="retro-input"
-                  placeholder="College Name"
-                />
-                {errors.collegeName && <div className="error-message">{errors.collegeName}</div>}
-              </div>
-
-              <div className="form-group">
-                <label className="form-label required">Department / Branch</label>
-                <input
-                  type="text"
-                  name="department"
-                  value={formData.department}
-                  onChange={handleInputChange}
-                  className="retro-input"
-                  placeholder="Department / Branch"
-                />
-                {errors.department && <div className="error-message">{errors.department}</div>}
-              </div>
-
-              <div className="form-group">
-                <label className="form-label required">Year of Study</label>
-                <input
-                  type="text"
-                  name="yearOfStudy"
-                  value={formData.yearOfStudy}
-                  onChange={handleInputChange}
-                  className="retro-input"
-                  placeholder="Year of Study"
-                />
-                {errors.yearOfStudy && <div className="error-message">{errors.yearOfStudy}</div>}
-              </div>
-
-              <div className="form-group">
-                <label className="form-label required">Contact Number</label>
-                <input
-                  type="text"
-                  name="contactNumber"
-                  value={formData.contactNumber}
-                  onChange={handleInputChange}
-                  className="retro-input"
-                  placeholder="Contact Number"
-                />
-                {errors.contactNumber && <div className="error-message">{errors.contactNumber}</div>}
-              </div>
-
-              <div className="form-group">
-                <label className="form-label required">Email Address</label>
-                <input
-                  type="text"
-                  name="emailAddress"
-                  value={formData.emailAddress}
-                  onChange={handleInputChange}
-                  className="retro-input"
-                  placeholder="Email Address"
-                />
-                {errors.emailAddress && <div className="error-message">{errors.emailAddress}</div>}
-              </div>
-
-              <div className="form-group">
-                <label className="form-label required">FIFA Mobile Username (In-game Name)</label>
-                <input
-                  type="text"
-                  name="fifaUsername"
-                  value={formData.fifaUsername}
-                  onChange={handleInputChange}
-                  className="retro-input"
-                  placeholder="FIFA Mobile Username"
-                />
-                {errors.fifaUsername && <div className="error-message">{errors.fifaUsername}</div>}
-              </div>
-
-              <div className="form-group">
-                <label className="form-label required">Team OVR (Overall Rating)</label>
-                <input
-                  type="text"
-                  name="teamOvr"
-                  value={formData.teamOvr}
-                  onChange={handleInputChange}
-                  className="retro-input"
-                  placeholder="Team OVR"
-                />
-                {errors.teamOvr && <div className="error-message">{errors.teamOvr}</div>}
-              </div>
-
-              <div className="form-group">
-                <label className="form-label required">Device Model</label>
-                <input
-                  type="text"
-                  name="deviceModel"
-                  value={formData.deviceModel}
-                  onChange={handleInputChange}
-                  className="retro-input"
-                  placeholder="Device Model"
-                />
-                {errors.deviceModel && <div className="error-message">{errors.deviceModel}</div>}
-              </div>
-            </div>
-
-            <div className="form-section">
-              <h2 className="form-section-title">&gt;&gt;&gt; Payment Information</h2>
-              
-              <div className="form-group">
-                <label className="form-label required">Payment Mode</label>
-                <div className="mcq-group">
-                  <label className="mcq-option">
-                    <input
-                      type="radio"
-                      name="paymentMode"
-                      value="online"
-                      checked={formData.paymentMode === 'online'}
-                      onChange={handleInputChange}
-                    />
-                    <span className="mcq-option-label">Online Payment</span>
-                  </label>
-                  <label className="mcq-option">
-                    <input
-                      type="radio"
-                      name="paymentMode"
-                      value="offline"
-                      checked={formData.paymentMode === 'offline'}
-                      onChange={handleInputChange}
-                    />
-                    <span className="mcq-option-label">Offline Payment</span>
-                  </label>
-                </div>
-                {errors.paymentMode && <div className="error-message">{errors.paymentMode}</div>}
-              </div>
-
-              <div className="form-group">
-                <label className="form-label required">Payment Date</label>
-                <input
-                  type="date"
-                  name="paymentDate"
-                  value={formData.paymentDate}
-                  onChange={handleInputChange}
-                  className="retro-input"
-                  max={new Date().toISOString().split('T')[0]}
-                />
-                {errors.paymentDate && <div className="error-message">{errors.paymentDate}</div>}
-              </div>
-
-              {formData.paymentMode === 'online' && (
                 <div className="form-group">
-                  <label className="form-label required">Transaction ID</label>
+                  <label className="form-label required">Full Name</label>
                   <input
                     type="text"
-                    name="transactionId"
-                    value={formData.transactionId}
+                    name="fullName"
+                    value={formData.fullName}
                     onChange={handleInputChange}
                     className="retro-input"
-                    placeholder="Transaction ID"
+                    placeholder="Full Name"
                   />
-                  {errors.transactionId && <div className="error-message">{errors.transactionId}</div>}
+                  {errors.fullName && <div className="error-message">{errors.fullName}</div>}
                 </div>
-              )}
 
-              {formData.paymentMode && (
                 <div className="form-group">
-                  <label className="form-label required">
-                    {formData.paymentMode === 'online' ? 'Upload Payment Screenshot' : 'Upload Offline Receipt'}
-                  </label>
+                  <label className="form-label required">College Name</label>
+                  <div className="mcq-group">
+                    {COLLEGE_OPTIONS.map((option) => (
+                      <label className="mcq-option" key={option}>
+                        <input
+                          type="radio"
+                          name="collegeName"
+                          value={option}
+                          checked={formData.collegeName === option}
+                          onChange={handleInputChange}
+                        />
+                        <span className="mcq-option-label">{option}</span>
+                      </label>
+                    ))}
+                  </div>
+                  {errors.collegeName && <div className="error-message">{errors.collegeName}</div>}
+                </div>
+
+                {formData.collegeName === 'Others' && (
+                  <div className="form-group">
+                    <label className="form-label required">Specify College Name</label>
+                    <input
+                      type="text"
+                      name="collegeOther"
+                      value={formData.collegeOther}
+                      onChange={handleInputChange}
+                      className="retro-input"
+                      placeholder="Enter your college name"
+                    />
+                    {errors.collegeOther && <div className="error-message">{errors.collegeOther}</div>}
+                  </div>
+                )}
+
+                <div className="form-group">
+                  <label className="form-label required">Department / Branch</label>
+                  <input
+                    type="text"
+                    name="department"
+                    value={formData.department}
+                    onChange={handleInputChange}
+                    className="retro-input"
+                    placeholder="Department / Branch"
+                  />
+                  {errors.department && <div className="error-message">{errors.department}</div>}
+                </div>
+
+                <div className="form-group">
+                  <label className="form-label required">Year of Study</label>
+                  <div className="mcq-group">
+                    {YEAR_OPTIONS.map((yearOption) => (
+                      <label className="mcq-option" key={yearOption}>
+                        <input
+                          type="radio"
+                          name="yearOfStudy"
+                          value={yearOption}
+                          checked={formData.yearOfStudy === yearOption}
+                          onChange={handleInputChange}
+                        />
+                        <span className="mcq-option-label">{yearOption}</span>
+                      </label>
+                    ))}
+                  </div>
+                  {errors.yearOfStudy && <div className="error-message">{errors.yearOfStudy}</div>}
+                </div>
+
+                <div className="form-group">
+                  <label className="form-label required">Contact Number</label>
+                  <input
+                    type="text"
+                    name="contactNumber"
+                    value={formData.contactNumber}
+                    onChange={handleInputChange}
+                    className="retro-input"
+                    placeholder="Contact Number"
+                  />
+                  {errors.contactNumber && <div className="error-message">{errors.contactNumber}</div>}
+                </div>
+
+                <div className="form-group">
+                  <label className="form-label required">Email Address</label>
+                  <input
+                    type="text"
+                    name="emailAddress"
+                    value={formData.emailAddress}
+                    onChange={handleInputChange}
+                    className="retro-input"
+                    placeholder="Email Address"
+                  />
+                  {errors.emailAddress && <div className="error-message">{errors.emailAddress}</div>}
+                </div>
+
+                <div className="form-group">
+                  <label className="form-label required">College ID / Library Card Upload</label>
                   <div className="file-upload-wrapper">
                     <div className="file-upload">
                       <input
                         type="file"
-                        name="paymentReceipt"
-                        id="paymentReceipt"
-                        onChange={handleInputChange}
+                        name="collegeIdProof"
+                        id="collegeIdProof"
                         className="file-upload-input"
                         accept="image/jpeg,image/jpg,image/png,.jpg,.jpeg,.png"
+                        onChange={handleInputChange}
                       />
-                      <label htmlFor="paymentReceipt" className="file-upload-label">
-                        <div className="file-upload-icon"></div>
+                      <label htmlFor="collegeIdProof" className="file-upload-label">
+                        <div className="file-upload-icon">📁</div>
                         <div className="file-upload-text">
-                          <span className="highlight">Click to upload</span><br/>
-                          {formData.paymentMode === 'online' ? 'Payment screenshot' : 'Offline receipt'}<br/>
-                          (PNG, JPG, JPEG)
+                          <span className="highlight">Click to upload</span>
+                          <br />
+                          PNG, JPG, JPEG
                         </div>
                       </label>
                     </div>
-                    {formData.paymentReceipt && (
-                      <div className="file-name">
-                        ✓ {formData.paymentReceipt.name}
-                      </div>
-                    )}
+                    {formData.collegeIdProof && <div className="file-name">✓ {formData.collegeIdProof.name}</div>}
                   </div>
-                  {errors.paymentReceipt && <div className="error-message">{errors.paymentReceipt}</div>}
+                  {errors.collegeIdProof && <div className="error-message">{errors.collegeIdProof}</div>}
                 </div>
-              )}
-            </div>
 
-            <div className="form-section">
-              <h2 className="form-section-title">&gt;&gt;&gt; Agreement</h2>
-
-              <div className="form-group">
-                <label className="form-label required">Agreement to Tournament Rules</label>
-                <label className="checkbox-group">
+                <div className="form-group">
+                  <label className="form-label required">FIFA Mobile Username (In-game Name)</label>
                   <input
-                    type="checkbox"
-                    name="agreeToRules"
-                    checked={formData.agreeToRules}
+                    type="text"
+                    name="fifaUsername"
+                    value={formData.fifaUsername}
                     onChange={handleInputChange}
+                    className="retro-input"
+                    placeholder="FIFA Mobile Username"
                   />
-                  <span className="checkbox-custom"></span>
-                  <span className="checkbox-label">
-                    I agree to the FIFA Mobile tournament rules
-                  </span>
-                </label>
-                {errors.agreeToRules && <div className="error-message">{errors.agreeToRules}</div>}
-              </div>
-            </div>
+                  {errors.fifaUsername && <div className="error-message">{errors.fifaUsername}</div>}
+                </div>
 
-            <div className="form-section">
-              <h2 className="form-section-title">&gt;&gt;&gt; WhatsApp Group</h2>
-              <div className="form-group">
-                <label className="form-label">Link</label>
-                <p style={{ margin: 0 }}>
-                  <a
-                    href="https://chat.whatsapp.com/"
-                    target="_blank"
-                    rel="noopener noreferrer"
-                    style={{ color: '#ffc010' }}
-                  >
-                    https://chat.whatsapp.com/
-                  </a>
-                </p>
-              </div>
-
-              <div className="form-group">
-                <label className="checkbox-group">
+                <div className="form-group">
+                  <label className="form-label required">Team OVR (Overall Rating)</label>
                   <input
-                    type="checkbox"
-                    name="whatsappConfirmed"
-                    checked={formData.whatsappConfirmed}
+                    type="text"
+                    name="teamOvr"
+                    value={formData.teamOvr}
                     onChange={handleInputChange}
+                    className="retro-input"
+                    placeholder="Team OVR"
                   />
-                  <span className="checkbox-custom"></span>
-                  <span className="checkbox-label">
-                    Yes, I have Done
-                  </span>
-                </label>
-                {errors.whatsappConfirmed && <div className="error-message">{errors.whatsappConfirmed}</div>}
-              </div>
-            </div>
+                  {errors.teamOvr && <div className="error-message">{errors.teamOvr}</div>}
+                </div>
 
-            <div className="submit-button-wrapper">
-              <button type="submit" className="retro-button" disabled={isSubmitting}>
-                {isSubmitting ? 'Submitting...' : 'Submit Registration'}
-              </button>
-              <button
-                type="button"
-                className="retro-button secondary"
-                onClick={() => history.goBack()}
-                disabled={isSubmitting}
-              >
-                Cancel
-              </button>
-            </div>
-
-            {submitSuccess && (
-              <div className="success-message" style={{ marginTop: '20px' }}>
-                ✓ Registration Successful! Redirecting to events page...
+                <div className="form-group">
+                  <label className="form-label required">Device Model</label>
+                  <input
+                    type="text"
+                    name="deviceModel"
+                    value={formData.deviceModel}
+                    onChange={handleInputChange}
+                    className="retro-input"
+                    placeholder="Device Model"
+                  />
+                  {errors.deviceModel && <div className="error-message">{errors.deviceModel}</div>}
+                </div>
               </div>
-            )}
-          </form>
+            </Step>
+
+            {/* Step 2: Payment Information */}
+            <Step>
+              <div className="form-section">
+                <h2 className="form-section-title">&gt;&gt;&gt; Payment Information</h2>
+
+                <div style={{
+                  backgroundColor: 'rgba(255, 192, 16, 0.1)',
+                  border: '2px solid #ffc010',
+                  padding: '20px',
+                  borderRadius: '5px',
+                  marginBottom: '25px'
+                }}>
+                  <h4 style={{ color: '#ffc010', marginTop: 0, marginBottom: '15px', fontSize: '14px' }}>
+                    💰 Registration Fee
+                  </h4>
+                  <p style={{ color: '#fff', lineHeight: '1.8', margin: 0, fontSize: '16px' }}>
+                    <strong style={{ color: '#ffc010' }}>In-house (BPPIMT):</strong> ₹50 per participant
+                    <br />
+                    <strong style={{ color: '#ffc010' }}>Outside College:</strong> ₹60 per participant
+                  </p>
+                </div>
+                
+                <div className="form-group">
+                  <label className="form-label required">Payment Mode</label>
+                  <div className="mcq-group">
+                    <label className="mcq-option">
+                      <input
+                        type="radio"
+                        name="paymentMode"
+                        value="online"
+                        checked={formData.paymentMode === 'online'}
+                        onChange={handleInputChange}
+                      />
+                      <span className="mcq-option-label">Online Payment</span>
+                    </label>
+                    <label className="mcq-option">
+                      <input
+                        type="radio"
+                        name="paymentMode"
+                        value="offline"
+                        checked={formData.paymentMode === 'offline'}
+                        onChange={handleInputChange}
+                      />
+                      <span className="mcq-option-label">Offline Payment</span>
+                    </label>
+                  </div>
+                  {errors.paymentMode && <div className="error-message">{errors.paymentMode}</div>}
+                </div>
+
+                {formData.paymentMode === 'online' && (
+                  <>
+                    {/* QR Code and UPI ID Section */}
+                    <div className="payment-qr-section">
+                      <div className="payment-qr-title">SCAN QR CODE TO PAY</div>
+                      <div className="payment-qr-container">
+                        <div className="qr-code-wrapper">
+                          <img 
+                            src={qrCodeImage} 
+                            alt="Payment QR Code" 
+                            className="qr-code-image"
+                          />
+                        </div>
+                        <div className="upi-id-container">
+                          <div className="upi-id-label">UPI ID</div>
+                          <div className="upi-id-value">bppoddar@iob</div>
+                        </div>
+                        <div className="payment-instruction-note">
+                          Scan the QR code or use the UPI ID to make payment
+                        </div>
+                      </div>
+                    </div>
+
+                    <div className="form-group">
+                      <label className="form-label required">Payment Date</label>
+                      <input
+                        type="date"
+                        name="paymentDate"
+                        value={formData.paymentDate}
+                        onChange={handleInputChange}
+                        className="retro-input"
+                        max={new Date().toISOString().split('T')[0]}
+                      />
+                      {errors.paymentDate && <div className="error-message">{errors.paymentDate}</div>}
+                    </div>
+
+                    <div className="form-group">
+                      <label className="form-label required">Transaction ID</label>
+                      <input
+                        type="text"
+                        name="transactionId"
+                        value={formData.transactionId}
+                        onChange={handleInputChange}
+                        className="retro-input"
+                        placeholder="Transaction ID"
+                      />
+                      {errors.transactionId && <div className="error-message">{errors.transactionId}</div>}
+                    </div>
+
+                    <div className="form-group">
+                      <label className="form-label required">Upload Payment Screenshot</label>
+                      <div className="file-upload-wrapper">
+                        <div className="file-upload">
+                          <input
+                            type="file"
+                            name="paymentReceipt"
+                            id="paymentReceipt"
+                            onChange={(e) => {
+                              const file = e.target.files[0];
+                              setFormData(prev => ({ ...prev, paymentReceipt: file }));
+                              if (errors.paymentReceipt) {
+                                setErrors(prev => ({ ...prev, paymentReceipt: '' }));
+                              }
+                            }}
+                            className="file-upload-input"
+                            accept="image/jpeg,image/jpg,image/png,.jpg,.jpeg,.png"
+                          />
+                          <label htmlFor="paymentReceipt" className="file-upload-label">
+                            <div className="file-upload-icon">💳</div>
+                            <div className="file-upload-text">
+                              <span className="highlight">Click to upload</span><br/>
+                              Payment screenshot<br/>
+                              (PNG, JPG, JPEG)
+                            </div>
+                          </label>
+                        </div>
+                        {formData.paymentReceipt && (
+                          <div className="file-name">
+                            ✓ {formData.paymentReceipt.name}
+                          </div>
+                        )}
+                      </div>
+                      {errors.paymentReceipt && <div className="error-message">{errors.paymentReceipt}</div>}
+                    </div>
+                  </>
+                )}
+
+                {formData.paymentMode === 'offline' && (
+                  <div className="form-group">
+                    <div style={{
+                      backgroundColor: 'rgba(255, 192, 16, 0.1)',
+                      border: '2px solid #ffc010',
+                      padding: '20px',
+                      borderRadius: '5px',
+                      marginTop: '20px'
+                    }}>
+                      <h4 style={{ color: '#ffc010', marginTop: 0, marginBottom: '15px' }}>
+                        📌 Important: Cash Payment Instructions
+                      </h4>
+                      <p style={{ color: '#fff', lineHeight: '1.8', margin: 0 }}>
+                        You must pay the registration fee in cash within <strong style={{ color: '#ffc010' }}>7 days of registration</strong> at the <strong style={{ color: '#ffc010' }}>Alumni Room</strong> of the college. 
+                        <br /><br />
+                        Please collect the paper receipt during payment and <strong style={{ color: '#ffc010' }}>keep it safe for event day verification</strong>.
+                        <br /><br />
+                        <span style={{ color: '#ff6b6b' }}>⚠️ No receipt upload is required during registration.</span>
+                      </p>
+                    </div>
+                  </div>
+                )}
+              </div>
+            </Step>
+
+            {/* Step 3: Rules & Agreement */}
+            <Step>
+              <div className="form-section">
+                <h2 className="form-section-title">&gt;&gt;&gt; Rules & Agreement</h2>
+
+                <div className="rules-container" style={{ 
+                  background: '#1a1a1a', 
+                  padding: '30px', 
+                  borderRadius: '8px',
+                  marginBottom: '30px',
+                  border: '2px solid #333'
+                }}>
+                  <h3 style={{ color: '#ffc010', marginBottom: '20px' }}>FIFA Mobile Tournament Rules</h3>
+                  <ul style={{ color: '#ccc', lineHeight: '1.8', paddingLeft: '20px' }}>
+                    <li>All participants must carry a valid college ID card</li>
+                    <li>Registration is mandatory for all participants</li>
+                    <li>Participants must follow the tournament schedule strictly</li>
+                    <li>Any form of cheating or malpractice will lead to disqualification</li>
+                    <li>Organizers' decision will be final and binding</li>
+                    <li>Entry fee is non-refundable under any circumstances</li>
+                  </ul>
+                </div>
+
+                <div className="form-group">
+                  <label className="form-label required">Agreement to Tournament Rules</label>
+                  <label className="checkbox-group">
+                    <input
+                      type="checkbox"
+                      name="agreeToRules"
+                      checked={formData.agreeToRules}
+                      onChange={handleInputChange}
+                    />
+                    <span className="checkbox-custom"></span>
+                    <span className="checkbox-label">
+                      I agree to the FIFA Mobile tournament rules
+                    </span>
+                  </label>
+                  {errors.agreeToRules && <div className="error-message">{errors.agreeToRules}</div>}
+                </div>
+
+                <div className="form-group">
+                  <h3 className="form-section-title">&gt;&gt;&gt; WhatsApp Group</h3>
+                  <label className="form-label">Link</label>
+                  <p style={{ margin: '10px 0' }}>
+                    <a
+                      href="https://chat.whatsapp.com/"
+                      target="_blank"
+                      rel="noopener noreferrer"
+                      style={{ color: '#ffc010' }}
+                    >
+                      https://chat.whatsapp.com/
+                    </a>
+                  </p>
+                </div>
+
+                <div className="form-group">
+                  <label className="checkbox-group">
+                    <input
+                      type="checkbox"
+                      name="whatsappConfirmed"
+                      checked={formData.whatsappConfirmed}
+                      onChange={handleInputChange}
+                    />
+                    <span className="checkbox-custom"></span>
+                    <span className="checkbox-label">
+                      Yes, I have joined the WhatsApp group
+                    </span>
+                  </label>
+                  {errors.whatsappConfirmed && <div className="error-message">{errors.whatsappConfirmed}</div>}
+                </div>
+              </div>
+            </Step>
+          </Stepper>
+
+          {submitSuccess && (
+            <div className="success-message" style={{ marginTop: '20px', textAlign: 'center', padding: '20px', background: '#1a1a1a', borderRadius: '8px' }}>
+              <h3 style={{ color: '#ffc010' }}>✓ Registration Successful!</h3>
+              <p style={{ color: '#999' }}>Redirecting to events page...</p>
+            </div>
+          )}
+
+          <div style={{ marginTop: '20px', textAlign: 'center' }}>
+            <button 
+              type="button" 
+              className="retro-button secondary"
+              onClick={() => history.goBack()}
+              disabled={isSubmitting}
+            >
+              Cancel Registration
+            </button>
+          </div>
         </div>
       </div>
     </div>
