@@ -2,8 +2,16 @@ import React, { useState } from 'react';
 import { useHistory } from 'react-router-dom';
 import { submitEventRegistration } from '../../../utils/eventRegistrationAPI';
 import Breadcrumb from '../../Utilities/Breadcrumb/Breadcrumb';
+import Stepper, { Step } from '../../Utilities/Stepper/Stepper';
 import './Registration.css';
 import forzaHorizonBanner from '../../../assets/img/event_specific_pictures/games/forza_horizon.png';
+import qrCodeImage from '../../../assets/img/QrCode_For_Payment.jpg.jpeg';
+
+const YEAR_OPTIONS = ['1st Year', '2nd Year', '3rd Year', '4th Year'];
+const COLLEGE_OPTIONS = [
+  'B. P. Poddar Institute of Management & Technology',
+  'Others'
+];
 
 const ForzaHorizonRegistration = () => {
   const history = useHistory();
@@ -13,6 +21,8 @@ const ForzaHorizonRegistration = () => {
     contactNumber: '',
     emailId: '',
     collegeName: '',
+    collegeOther: '',
+    year: '',
     participantIdProof: null,
     paymentMode: '',
     transactionId: '',
@@ -40,28 +50,33 @@ const ForzaHorizonRegistration = () => {
     }
   };
 
-  const validateForm = () => {
+  const validateStep1 = () => {
     const nextErrors = {};
-
     if (!formData.participantName.trim()) nextErrors.participantName = 'Participant Name is required';
     if (!formData.name.trim()) nextErrors.name = 'Name is required';
-
     if (!formData.contactNumber.trim()) {
       nextErrors.contactNumber = 'Contact Number is required';
     } else if (!/^\d{10,15}$/.test(formData.contactNumber.replace(/\D/g, ''))) {
       nextErrors.contactNumber = 'Enter a valid contact number';
     }
-
     if (!formData.emailId.trim()) {
       nextErrors.emailId = 'Email ID is required';
     } else if (!/\S+@\S+\.\S+/.test(formData.emailId)) {
       nextErrors.emailId = 'Invalid email format';
     }
-
-    if (!formData.collegeName.trim()) nextErrors.collegeName = 'College Name is required';
+    if (!formData.collegeName) nextErrors.collegeName = 'College selection is required';
+    if (formData.collegeName === 'Others' && !formData.collegeOther.trim()) {
+      nextErrors.collegeOther = 'Please specify your college name';
+    }
+    if (!formData.year) nextErrors.year = 'Year is required';
     if (!formData.participantIdProof) nextErrors.participantIdProof = 'Participant ID Proof is required';
-    if (!formData.paymentMode) nextErrors.paymentMode = 'Mode of Payment is required';
+    setErrors(nextErrors);
+    return Object.keys(nextErrors).length === 0;
+  };
 
+  const validateStep2 = () => {
+    const nextErrors = {};
+    if (!formData.paymentMode) nextErrors.paymentMode = 'Mode of Payment is required';
     if (formData.paymentMode === 'online') {
       if (!formData.transactionId.trim()) {
         nextErrors.transactionId = 'Transaction ID is required for online payment';
@@ -70,22 +85,31 @@ const ForzaHorizonRegistration = () => {
         nextErrors.paymentScreenshot = 'Upload payment screenshot for online payment';
       }
     }
-
-    if (!formData.whatsappConfirmed) {
-      nextErrors.whatsappConfirmed = 'Please confirm after joining the WhatsApp group';
-    }
-
     setErrors(nextErrors);
     return Object.keys(nextErrors).length === 0;
   };
 
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-
-    if (!validateForm()) {
-      return;
+  const validateStep3 = () => {
+    const nextErrors = {};
+    if (!formData.whatsappConfirmed) {
+      nextErrors.whatsappConfirmed = 'Please confirm after joining the WhatsApp group';
     }
+    setErrors(nextErrors);
+    return Object.keys(nextErrors).length === 0;
+  };
 
+  const handleStepValidation = (step) => {
+    if (step === 1) {
+      return validateStep1();
+    } else if (step === 2) {
+      return validateStep2();
+    } else if (step === 3) {
+      return validateStep3();
+    }
+    return true;
+  };
+
+  const handleComplete = async () => {
     setIsSubmitting(true);
 
     try {
@@ -134,232 +158,354 @@ const ForzaHorizonRegistration = () => {
             </div>
           )}
 
-          <form className="registration-form" onSubmit={handleSubmit}>
-            <div className="form-section">
-              <h2 className="form-section-title">&gt;&gt;&gt; Participant Details</h2>
+          <Stepper
+            initialStep={1}
+            onStepChange={(step) => console.log('Step:', step)}
+            onStepValidation={handleStepValidation}
+            onFinalStepCompleted={handleComplete}
+            backButtonText="← Previous"
+            nextButtonText="Next Step →"
+          >
+            <Step>
+              <div className="form-section">
+                <h2 className="form-section-title">&gt;&gt;&gt; Participant Details</h2>
 
-              <div className="form-group">
-                <label className="form-label required">Participant Name</label>
-                <input
-                  type="text"
-                  name="participantName"
-                  value={formData.participantName}
-                  onChange={handleInputChange}
-                  className="retro-input"
-                  placeholder="Participant Name"
-                />
-                {errors.participantName && <div className="error-message">{errors.participantName}</div>}
-              </div>
-            </div>
-
-            <div className="form-section">
-              <h2 className="form-section-title">&gt;&gt;&gt; Participant (Mandatory)</h2>
-
-              <div className="form-group">
-                <label className="form-label required">Name</label>
-                <input
-                  type="text"
-                  name="name"
-                  value={formData.name}
-                  onChange={handleInputChange}
-                  className="retro-input"
-                  placeholder="Name"
-                />
-                {errors.name && <div className="error-message">{errors.name}</div>}
-              </div>
-
-              <div className="form-group">
-                <label className="form-label required">Contact Number</label>
-                <input
-                  type="text"
-                  name="contactNumber"
-                  value={formData.contactNumber}
-                  onChange={handleInputChange}
-                  className="retro-input"
-                  placeholder="Contact Number"
-                />
-                {errors.contactNumber && <div className="error-message">{errors.contactNumber}</div>}
-              </div>
-
-              <div className="form-group">
-                <label className="form-label required">Email ID</label>
-                <input
-                  type="text"
-                  name="emailId"
-                  value={formData.emailId}
-                  onChange={handleInputChange}
-                  className="retro-input"
-                  placeholder="Email ID"
-                />
-                {errors.emailId && <div className="error-message">{errors.emailId}</div>}
-              </div>
-
-              <div className="form-group">
-                <label className="form-label required">College Name</label>
-                <input
-                  type="text"
-                  name="collegeName"
-                  value={formData.collegeName}
-                  onChange={handleInputChange}
-                  className="retro-input"
-                  placeholder="College Name"
-                />
-                {errors.collegeName && <div className="error-message">{errors.collegeName}</div>}
-              </div>
-
-              <div className="form-group">
-                <label className="form-label required">Participant ID Proof (College ID / Library Card)</label>
-                <div className="file-upload-wrapper">
-                  <div className="file-upload">
-                    <input
-                      type="file"
-                      name="participantIdProof"
-                      id="participantIdProof"
-                      className="file-upload-input"
-                      accept="image/jpeg,image/jpg,image/png,.jpg,.jpeg,.png"
-                      onChange={handleInputChange}
-                    />
-                    <label htmlFor="participantIdProof" className="file-upload-label">
-                      <div className="file-upload-icon"></div>
-                      <div className="file-upload-text">
-                        <span className="highlight">Click to upload</span>
-                        <br />
-                        PNG, JPG, JPEG
-                      </div>
-                    </label>
-                  </div>
-                  {formData.participantIdProof && <div className="file-name">{formData.participantIdProof.name}</div>}
-                </div>
-                {errors.participantIdProof && <div className="error-message">{errors.participantIdProof}</div>}
-              </div>
-            </div>
-
-            <div className="form-section">
-              <h2 className="form-section-title">&gt;&gt;&gt; Payment Details</h2>
-
-              <div className="form-group">
-                <label className="form-label required">Mode of Payment</label>
-                <div className="mcq-group">
-                  <label className="mcq-option">
-                    <input
-                      type="radio"
-                      name="paymentMode"
-                      value="cash"
-                      checked={formData.paymentMode === 'cash'}
-                      onChange={handleInputChange}
-                    />
-                    <span className="mcq-option-label">Cash</span>
-                  </label>
-
-                  <label className="mcq-option">
-                    <input
-                      type="radio"
-                      name="paymentMode"
-                      value="online"
-                      checked={formData.paymentMode === 'online'}
-                      onChange={handleInputChange}
-                    />
-                    <span className="mcq-option-label">Online</span>
-                  </label>
-                </div>
-                {errors.paymentMode && <div className="error-message">{errors.paymentMode}</div>}
-              </div>
-
-              <div className="form-group">
-                <label className="form-label">Transaction ID</label>
-                <input
-                  type="text"
-                  name="transactionId"
-                  value={formData.transactionId}
-                  onChange={handleInputChange}
-                  className="retro-input"
-                  placeholder="Transaction ID"
-                />
-                {errors.transactionId && <div className="error-message">{errors.transactionId}</div>}
-              </div>
-
-              <div className="form-group">
-                <label className="form-label">Upload Payment Screenshot (For Online Payment)</label>
-                <div className="file-upload-wrapper">
-                  <div className="file-upload">
-                    <input
-                      type="file"
-                      name="paymentScreenshot"
-                      id="paymentScreenshot"
-                      className="file-upload-input"
-                      accept="image/jpeg,image/jpg,image/png,.jpg,.jpeg,.png"
-                      onChange={handleInputChange}
-                    />
-                    <label htmlFor="paymentScreenshot" className="file-upload-label">
-                      <div className="file-upload-icon"></div>
-                      <div className="file-upload-text">
-                        <span className="highlight">Click to upload</span>
-                        <br />
-                        PNG, JPG, JPEG
-                      </div>
-                    </label>
-                  </div>
-                  {formData.paymentScreenshot && <div className="file-name">{formData.paymentScreenshot.name}</div>}
-                </div>
-                {errors.paymentScreenshot && <div className="error-message">{errors.paymentScreenshot}</div>}
-              </div>
-
-              <div className="form-group">
-                <label className="form-label">Registration fee</label>
-                <p style={{ margin: 0, color: '#ffffff', fontFamily: 'Press Start 2P, monospace', fontSize: '10px', lineHeight: '1.8' }}>
-                  In house: 80
-                  <br />
-                  Outer college: 100
-                </p>
-              </div>
-            </div>
-
-            <div className="form-section">
-              <h2 className="form-section-title">&gt;&gt;&gt; Communication</h2>
-              <div className="form-group">
-                <label className="form-label required">Join WhatsApp Group</label>
-                <p style={{ margin: 0, color: '#ffc010', fontFamily: 'Press Start 2P, monospace', fontSize: '10px' }}>
-                  Link will be provided by organizers
-                </p>
-              </div>
-
-              <div className="form-group">
-                <label className="checkbox-group">
+                <div className="form-group">
+                  <label className="form-label required">Participant Name</label>
                   <input
-                    type="checkbox"
-                    name="whatsappConfirmed"
-                    checked={formData.whatsappConfirmed}
+                    type="text"
+                    name="participantName"
+                    value={formData.participantName}
                     onChange={handleInputChange}
+                    className="retro-input"
+                    placeholder="Participant Name"
                   />
-                  <span className="checkbox-custom"></span>
-                  <span className="checkbox-label">
-                    I confirm that I have checked all details carefully and joined the WhatsApp group
-                  </span>
-                </label>
-                {errors.whatsappConfirmed && <div className="error-message">{errors.whatsappConfirmed}</div>}
-              </div>
-            </div>
+                  {errors.participantName && <div className="error-message">{errors.participantName}</div>}
+                </div>
 
-            <div className="submit-button-wrapper">
-              <button type="submit" className="retro-button" disabled={isSubmitting}>
-                {isSubmitting ? 'Submitting...' : 'Submit Registration'}
-              </button>
-              <button
-                type="button"
-                className="retro-button secondary"
-                onClick={() => history.goBack()}
-                disabled={isSubmitting}
-              >
-                Cancel
-              </button>
-            </div>
+                <div className="form-group">
+                  <label className="form-label required">Name</label>
+                  <input
+                    type="text"
+                    name="name"
+                    value={formData.name}
+                    onChange={handleInputChange}
+                    className="retro-input"
+                    placeholder="Name"
+                  />
+                  {errors.name && <div className="error-message">{errors.name}</div>}
+                </div>
 
-            {submitSuccess && (
-              <div className="success-message" style={{ marginTop: '20px' }}>
-                ✓ Registration Successful! Redirecting to events page...
+                <div className="form-group">
+                  <label className="form-label required">Contact Number</label>
+                  <input
+                    type="text"
+                    name="contactNumber"
+                    value={formData.contactNumber}
+                    onChange={handleInputChange}
+                    className="retro-input"
+                    placeholder="Contact Number"
+                  />
+                  {errors.contactNumber && <div className="error-message">{errors.contactNumber}</div>}
+                </div>
+
+                <div className="form-group">
+                  <label className="form-label required">Email ID</label>
+                  <input
+                    type="text"
+                    name="emailId"
+                    value={formData.emailId}
+                    onChange={handleInputChange}
+                    className="retro-input"
+                    placeholder="Email ID"
+                  />
+                  {errors.emailId && <div className="error-message">{errors.emailId}</div>}
+                </div>
+
+                <div className="form-group">
+                  <label className="form-label required">College Name</label>
+                  <div className="mcq-group">
+                    {COLLEGE_OPTIONS.map((option) => (
+                      <label className="mcq-option" key={option}>
+                        <input
+                          type="radio"
+                          name="collegeName"
+                          value={option}
+                          checked={formData.collegeName === option}
+                          onChange={handleInputChange}
+                        />
+                        <span className="mcq-option-label">{option}</span>
+                      </label>
+                    ))}
+                  </div>
+                  {errors.collegeName && <div className="error-message">{errors.collegeName}</div>}
+                </div>
+
+                {formData.collegeName === 'Others' && (
+                  <div className="form-group">
+                    <label className="form-label required">Specify College Name</label>
+                    <input
+                      type="text"
+                      name="collegeOther"
+                      value={formData.collegeOther}
+                      onChange={handleInputChange}
+                      className="retro-input"
+                      placeholder="Enter your college name"
+                    />
+                    {errors.collegeOther && <div className="error-message">{errors.collegeOther}</div>}
+                  </div>
+                )}
+
+                <div className="form-group">
+                  <label className="form-label required">Year</label>
+                  <div className="mcq-group">
+                    {YEAR_OPTIONS.map((yearOption) => (
+                      <label className="mcq-option" key={yearOption}>
+                        <input
+                          type="radio"
+                          name="year"
+                          value={yearOption}
+                          checked={formData.year === yearOption}
+                          onChange={handleInputChange}
+                        />
+                        <span className="mcq-option-label">{yearOption}</span>
+                      </label>
+                    ))}
+                  </div>
+                  {errors.year && <div className="error-message">{errors.year}</div>}
+                </div>
+
+                <div className="form-group">
+                  <label className="form-label required">Participant ID Proof (College ID / Library Card)</label>
+                  <div className="file-upload-wrapper">
+                    <div className="file-upload">
+                      <input
+                        type="file"
+                        name="participantIdProof"
+                        id="participantIdProof"
+                        className="file-upload-input"
+                        accept="image/jpeg,image/jpg,image/png,.jpg,.jpeg,.png"
+                        onChange={handleInputChange}
+                      />
+                      <label htmlFor="participantIdProof" className="file-upload-label">
+                        <div className="file-upload-icon">📁</div>
+                        <div className="file-upload-text">
+                          <span className="highlight">Click to upload</span>
+                          <br />
+                          PNG, JPG, JPEG
+                        </div>
+                      </label>
+                    </div>
+                    {formData.participantIdProof && <div className="file-name">✓ {formData.participantIdProof.name}</div>}
+                  </div>
+                  {errors.participantIdProof && <div className="error-message">{errors.participantIdProof}</div>}
+                </div>
               </div>
-            )}
-          </form>
+            </Step>
+
+            <Step>
+              <div className="form-section">
+                <h2 className="form-section-title">&gt;&gt;&gt; Payment Details</h2>
+
+                <div style={{
+                  backgroundColor: 'rgba(255, 192, 16, 0.1)',
+                  border: '2px solid #ffc010',
+                  padding: '20px',
+                  borderRadius: '5px',
+                  marginBottom: '25px'
+                }}>
+                  <h4 style={{ color: '#ffc010', marginTop: 0, marginBottom: '15px', fontSize: '14px' }}>
+                    💰 Registration Fee
+                  </h4>
+                  <p style={{ color: '#fff', lineHeight: '1.8', margin: 0, fontSize: '16px' }}>
+                    <strong style={{ color: '#ffc010' }}>In-house (BPPIMT):</strong> ₹80 per participant
+                    <br />
+                    <strong style={{ color: '#ffc010' }}>Outside College:</strong> ₹100 per participant
+                  </p>
+                </div>
+
+                <div className="form-group">
+                  <label className="form-label required">Mode of Payment</label>
+                  <div className="mcq-group">
+                    <label className="mcq-option">
+                      <input
+                        type="radio"
+                        name="paymentMode"
+                        value="cash"
+                        checked={formData.paymentMode === 'cash'}
+                        onChange={handleInputChange}
+                      />
+                      <span className="mcq-option-label">Cash</span>
+                    </label>
+
+                    <label className="mcq-option">
+                      <input
+                        type="radio"
+                        name="paymentMode"
+                        value="online"
+                        checked={formData.paymentMode === 'online'}
+                        onChange={handleInputChange}
+                      />
+                      <span className="mcq-option-label">Online</span>
+                    </label>
+                  </div>
+                  {errors.paymentMode && <div className="error-message">{errors.paymentMode}</div>}
+                </div>
+
+                {formData.paymentMode === 'online' && (
+                  <>
+                    {/* QR Code and UPI ID Section */}
+                    <div className="payment-qr-section">
+                      <div className="payment-qr-title">SCAN QR CODE TO PAY</div>
+                      <div className="payment-qr-container">
+                        <div className="qr-code-wrapper">
+                          <img 
+                            src={qrCodeImage} 
+                            alt="Payment QR Code" 
+                            className="qr-code-image"
+                          />
+                        </div>
+                        <div className="upi-id-container">
+                          <div className="upi-id-label">UPI ID</div>
+                          <div className="upi-id-value">bppoddar@iob</div>
+                        </div>
+                        <div className="payment-instruction-note">
+                          Scan the QR code or use the UPI ID to make payment
+                        </div>
+                      </div>
+                    </div>
+
+                    <div className="form-group">
+                      <label className="form-label required">Transaction ID</label>
+                      <input
+                        type="text"
+                        name="transactionId"
+                        value={formData.transactionId}
+                        onChange={handleInputChange}
+                        className="retro-input"
+                        placeholder="Transaction ID"
+                      />
+                      {errors.transactionId && <div className="error-message">{errors.transactionId}</div>}
+                    </div>
+
+                    <div className="form-group">
+                      <label className="form-label required">Upload Payment Screenshot</label>
+                      <div className="file-upload-wrapper">
+                        <div className="file-upload">
+                          <input
+                            type="file"
+                            name="paymentScreenshot"
+                            id="paymentScreenshot"
+                            className="file-upload-input"
+                            accept="image/jpeg,image/jpg,image/png,.jpg,.jpeg,.png"
+                            onChange={handleInputChange}
+                          />
+                          <label htmlFor="paymentScreenshot" className="file-upload-label">
+                            <div className="file-upload-icon">💳</div>
+                            <div className="file-upload-text">
+                              <span className="highlight">Click to upload</span>
+                              <br />
+                              PNG, JPG, JPEG
+                            </div>
+                          </label>
+                        </div>
+                        {formData.paymentScreenshot && <div className="file-name">✓ {formData.paymentScreenshot.name}</div>}
+                      </div>
+                      {errors.paymentScreenshot && <div className="error-message">{errors.paymentScreenshot}</div>}
+                    </div>
+                  </>
+                )}
+
+                {formData.paymentMode === 'cash' && (
+                  <div className="form-group">
+                    <div style={{
+                      backgroundColor: 'rgba(255, 192, 16, 0.1)',
+                      border: '2px solid #ffc010',
+                      padding: '20px',
+                      borderRadius: '5px',
+                      marginTop: '20px'
+                    }}>
+                      <h4 style={{ color: '#ffc010', marginTop: 0, marginBottom: '15px' }}>
+                        📌 Important: Cash Payment Instructions
+                      </h4>
+                      <p style={{ color: '#fff', lineHeight: '1.8', margin: 0 }}>
+                        You must pay the registration fee in cash within <strong style={{ color: '#ffc010' }}>7 days of registration</strong> at the <strong style={{ color: '#ffc010' }}>Alumni Room</strong> of the college. 
+                        <br /><br />
+                        Please collect the paper receipt during payment and <strong style={{ color: '#ffc010' }}>keep it safe for event day verification</strong>.
+                        <br /><br />
+                        <span style={{ color: '#ff6b6b' }}>⚠️ No receipt upload is required during registration.</span>
+                      </p>
+                    </div>
+                  </div>
+                )}
+              </div>
+            </Step>
+
+            <Step>
+              <div className="form-section">
+                <h2 className="form-section-title">&gt;&gt;&gt; Rules & Communication</h2>
+                
+                <div className="rules-container" style={{ 
+                  background: '#1a1a1a', 
+                  padding: '30px', 
+                  borderRadius: '8px',
+                  marginBottom: '30px',
+                  border: '2px solid #333'
+                }}>
+                  <h3 style={{ color: '#ffc010', marginBottom: '20px' }}>Forza Horizon Event Rules</h3>
+                  <ul style={{ color: '#ccc', lineHeight: '1.8', paddingLeft: '20px' }}>
+                    <li>All participants must carry a valid college ID card</li>
+                    <li>Registration is mandatory for all participants</li>
+                    <li>Participants must follow the event schedule strictly</li>
+                    <li>Any form of malpractice will lead to disqualification</li>
+                    <li>Organizers' decision will be final and binding</li>
+                    <li>Entry fee is non-refundable under any circumstances</li>
+                  </ul>
+                </div>
+
+                <div className="form-group">
+                  <label className="form-label required">Join WhatsApp Group</label>
+                  <p style={{ margin: '10px 0', color: '#ffc010', fontFamily: 'Press Start 2P, monospace', fontSize: '10px' }}>
+                    Link will be provided by organizers
+                  </p>
+                </div>
+
+                <div className="form-group">
+                  <label className="checkbox-group">
+                    <input
+                      type="checkbox"
+                      name="whatsappConfirmed"
+                      checked={formData.whatsappConfirmed}
+                      onChange={handleInputChange}
+                    />
+                    <span className="checkbox-custom"></span>
+                    <span className="checkbox-label">
+                      I confirm that I have checked all details carefully and joined the WhatsApp group
+                    </span>
+                  </label>
+                  {errors.whatsappConfirmed && <div className="error-message">{errors.whatsappConfirmed}</div>}
+                </div>
+              </div>
+            </Step>
+          </Stepper>
+
+          {submitSuccess && (
+            <div className="success-message" style={{ marginTop: '20px', textAlign: 'center', padding: '20px', background: '#1a1a1a', borderRadius: '8px' }}>
+              <h3 style={{ color: '#ffc010' }}>✓ Registration Successful!</h3>
+              <p style={{ color: '#999' }}>Redirecting to events page...</p>
+            </div>
+          )}
+
+          <div style={{ marginTop: '20px', textAlign: 'center' }}>
+            <button 
+              type="button" 
+              className="retro-button secondary"
+              onClick={() => history.goBack()}
+              disabled={isSubmitting}
+            >
+              Cancel Registration
+            </button>
+          </div>
         </div>
       </div>
     </div>
