@@ -2,11 +2,17 @@ import React, { useState } from 'react';
 import { useHistory } from 'react-router-dom';
 import { submitEventRegistration } from '../../../utils/eventRegistrationAPI';
 import Breadcrumb from '../../Utilities/Breadcrumb/Breadcrumb';
+import Stepper, { Step } from '../../Utilities/Stepper/Stepper';
 import './Registration.css';
 import khetBanner from '../../../assets/img/event_specific_pictures/games/khet.png';
+import qrCodeImage from '../../../assets/img/QrCode_For_Payment.jpg.jpeg';
 
 const YEAR_OPTIONS = ['1st Year', '2nd Year', '3rd Year', '4th Year'];
 const STREAM_OPTIONS = ['CSE', 'IT', 'ECE', 'EE', 'Others'];
+const COLLEGE_OPTIONS = [
+  'B. P. Poddar Institute of Management & Technology',
+  'Others'
+];
 
 const KhetRegistration = () => {
   const history = useHistory();
@@ -17,6 +23,7 @@ const KhetRegistration = () => {
     year: '',
     streamBranch: '',
     collegeName: '',
+    collegeOther: '',
     collegeIdProof: null,
     contactNumber: '',
     emailAddress: '',
@@ -25,7 +32,6 @@ const KhetRegistration = () => {
     transactionId: '',
     transactionDate: '',
     paymentScreenshot: null,
-    cashReceipt: null,
     declarationConfirmed: false
   });
   const [errors, setErrors] = useState({});
@@ -48,44 +54,40 @@ const KhetRegistration = () => {
     }
   };
 
-  const validateForm = () => {
+  const validateStep1 = () => {
     const nextErrors = {};
-
-    // teamName is optional for individual events
-    // if (!formData.teamName.trim()) nextErrors.teamName = 'Team Name is required';
-
     if (!String(formData.numberOfParticipants).trim()) {
       nextErrors.numberOfParticipants = 'Number of Participants is required';
     } else if (Number(formData.numberOfParticipants) !== 1) {
       nextErrors.numberOfParticipants = 'This is an individual event. Number of Participants must be 1';
     }
-
     if (!formData.fullName.trim()) nextErrors.fullName = 'Full Name is required';
     if (!formData.year) nextErrors.year = 'Year is required';
     if (!formData.streamBranch) nextErrors.streamBranch = 'Stream / Branch is required';
-    if (!formData.collegeName.trim()) nextErrors.collegeName = 'College Name is required';
+    if (!formData.collegeName) nextErrors.collegeName = 'College selection is required';
+    if (formData.collegeName === 'Others' && !formData.collegeOther.trim()) {
+      nextErrors.collegeOther = 'Please specify your college name';
+    }
     if (!formData.collegeIdProof) nextErrors.collegeIdProof = 'College ID / Library Card upload is required';
-
     if (!formData.contactNumber.trim()) {
       nextErrors.contactNumber = 'Contact Number is required';
     } else if (!/^\d{10,15}$/.test(formData.contactNumber.replace(/\D/g, ''))) {
       nextErrors.contactNumber = 'Enter a valid contact number';
     }
-
     if (!formData.emailAddress.trim()) {
       nextErrors.emailAddress = 'Email Address is required';
     } else if (!/\S+@\S+\.\S+/.test(formData.emailAddress)) {
       nextErrors.emailAddress = 'Invalid email format';
     }
+    setErrors(nextErrors);
+    return Object.keys(nextErrors).length === 0;
+  };
 
-    if (!formData.joinWhatsappConfirmed) {
-      nextErrors.joinWhatsappConfirmed = 'Please confirm you joined the official WhatsApp group';
-    }
-
+  const validateStep2 = () => {
+    const nextErrors = {};
     if (!formData.paymentMode) nextErrors.paymentMode = 'Mode of Payment is required';
-    if (!formData.transactionDate) nextErrors.transactionDate = 'Payment date is required';
-
     if (formData.paymentMode === 'online') {
+      if (!formData.transactionDate) nextErrors.transactionDate = 'Payment date is required';
       if (!formData.transactionId.trim()) {
         nextErrors.transactionId = 'Transaction ID is required for online payment';
       }
@@ -93,31 +95,34 @@ const KhetRegistration = () => {
         nextErrors.paymentScreenshot = 'Upload payment screenshot for online payment';
       }
     }
-
-    if (formData.paymentMode === 'cash') {
-      if (!formData.cashReceipt) {
-        nextErrors.cashReceipt = 'Upload cash receipt / slip for offline payment';
-      }
-    }
-
-    if (!formData.declarationConfirmed) {
-      nextErrors.declarationConfirmed = 'Please confirm the mandatory declaration';
-    }
-
     setErrors(nextErrors);
     return Object.keys(nextErrors).length === 0;
   };
 
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-    console.log('🚀 Form submitted!', formData);
-
-    if (!validateForm()) {
-      console.log('❌ Validation failed', errors);
-      return;
+  const validateStep3 = () => {
+    const nextErrors = {};
+    if (!formData.joinWhatsappConfirmed) {
+      nextErrors.joinWhatsappConfirmed = 'Please confirm you joined the official WhatsApp group';
     }
+    if (!formData.declarationConfirmed) {
+      nextErrors.declarationConfirmed = 'Please confirm the mandatory declaration';
+    }
+    setErrors(nextErrors);
+    return Object.keys(nextErrors).length === 0;
+  };
 
-    console.log('✅ Validation passed, submitting...');
+  const handleStepValidation = (step) => {
+    if (step === 1) {
+      return validateStep1();
+    } else if (step === 2) {
+      return validateStep2();
+    } else if (step === 3) {
+      return validateStep3();
+    }
+    return true;
+  };
+
+  const handleComplete = async () => {
     setIsSubmitting(true);
 
     try {
@@ -166,254 +171,118 @@ const KhetRegistration = () => {
             </div>
           )}
 
-          <form className="registration-form" onSubmit={handleSubmit}>
+          <Stepper
+            initialStep={1}
+            onStepChange={(step) => console.log('Step:', step)}
+            onStepValidation={handleStepValidation}
+            onFinalStepCompleted={handleComplete}
+            backButtonText="← Previous"
+            nextButtonText="Next Step →"
+          >
+            {/* Step 1: Participant Details */}
+            <Step>
+              <div className="form-section">
+                <h2 className="form-section-title">&gt;&gt;&gt; Participant Details</h2>
 
-            <div className="form-section">
-              <h2 className="form-section-title">&gt;&gt;&gt; Participant Details</h2>
-
-              <div className="form-group">
-                <label className="form-label required">Full Name</label>
-                <input
-                  type="text"
-                  name="fullName"
-                  value={formData.fullName}
-                  onChange={handleInputChange}
-                  className="retro-input"
-                  placeholder="Full Name"
-                />
-                {errors.fullName && <div className="error-message">{errors.fullName}</div>}
-              </div>
-
-              <div className="form-group">
-                <label className="form-label required">Year</label>
-                <div className="mcq-group">
-                  {YEAR_OPTIONS.map((option) => (
-                    <label className="mcq-option" key={option}>
-                      <input
-                        type="radio"
-                        name="year"
-                        value={option}
-                        checked={formData.year === option}
-                        onChange={handleInputChange}
-                      />
-                      <span className="mcq-option-label">{option}</span>
-                    </label>
-                  ))}
-                </div>
-                {errors.year && <div className="error-message">{errors.year}</div>}
-              </div>
-
-              <div className="form-group">
-                <label className="form-label required">Stream / Branch</label>
-                <div className="mcq-group">
-                  {STREAM_OPTIONS.map((option) => (
-                    <label className="mcq-option" key={option}>
-                      <input
-                        type="radio"
-                        name="streamBranch"
-                        value={option}
-                        checked={formData.streamBranch === option}
-                        onChange={handleInputChange}
-                      />
-                      <span className="mcq-option-label">{option}</span>
-                    </label>
-                  ))}
-                </div>
-                {errors.streamBranch && <div className="error-message">{errors.streamBranch}</div>}
-              </div>
-
-              <div className="form-group">
-                <label className="form-label required">College Name</label>
-                <input
-                  type="text"
-                  name="collegeName"
-                  value={formData.collegeName}
-                  onChange={handleInputChange}
-                  className="retro-input"
-                  placeholder="College Name"
-                />
-                {errors.collegeName && <div className="error-message">{errors.collegeName}</div>}
-              </div>
-
-              <div className="form-group">
-                <label className="form-label required">College ID / Library Card Upload</label>
-                <div className="file-upload-wrapper">
-                  <div className="file-upload">
-                    <input
-                      type="file"
-                      name="collegeIdProof"
-                      id="collegeIdProof"
-                      className="file-upload-input"
-                      accept="image/jpeg,image/jpg,image/png,.jpg,.jpeg,.png"
-                      onChange={handleInputChange}
-                    />
-                    <label htmlFor="collegeIdProof" className="file-upload-label">
-                      <div className="file-upload-icon"></div>
-                      <div className="file-upload-text">
-                        <span className="highlight">Click to upload</span>
-                        <br />
-                        PNG, JPG, JPEG
-                      </div>
-                    </label>
-                  </div>
-                  {formData.collegeIdProof && <div className="file-name">{formData.collegeIdProof.name}</div>}
-                </div>
-                {errors.collegeIdProof && <div className="error-message">{errors.collegeIdProof}</div>}
-              </div>
-
-              <div className="form-group">
-                <label className="form-label required">Contact Number (Preferably WhatsApp Number)</label>
-                <input
-                  type="text"
-                  name="contactNumber"
-                  value={formData.contactNumber}
-                  onChange={handleInputChange}
-                  className="retro-input"
-                  placeholder="Contact Number"
-                />
-                {errors.contactNumber && <div className="error-message">{errors.contactNumber}</div>}
-              </div>
-
-              <div className="form-group">
-                <label className="form-label required">Email Address</label>
-                <input
-                  type="text"
-                  name="emailAddress"
-                  value={formData.emailAddress}
-                  onChange={handleInputChange}
-                  className="retro-input"
-                  placeholder="Email Address"
-                />
-                {errors.emailAddress && <div className="error-message">{errors.emailAddress}</div>}
-              </div>
-
-              <div className="form-group">
-                <label className="form-label required">Join Official WhatsApp Group</label>
-                <p style={{ margin: 0 }}>
-                  <a
-                    href="https://chat.whatsapp.com/LeNZPNivID87ydepOLpAvY?mode=gi_t"
-                    target="_blank"
-                    rel="noopener noreferrer"
-                    style={{ color: '#ffc010' }}
-                  >
-                    https://chat.whatsapp.com/LeNZPNivID87ydepOLpAvY?mode=gi_t
-                  </a>
-                </p>
-                <label className="checkbox-group" style={{ marginTop: '12px' }}>
+                <div className="form-group">
+                  <label className="form-label required">Full Name</label>
                   <input
-                    type="checkbox"
-                    name="joinWhatsappConfirmed"
-                    checked={formData.joinWhatsappConfirmed}
+                    type="text"
+                    name="fullName"
+                    value={formData.fullName}
                     onChange={handleInputChange}
+                    className="retro-input"
+                    placeholder="Full Name"
                   />
-                  <span className="checkbox-custom"></span>
-                  <span className="checkbox-label">I have joined the official WhatsApp group</span>
-                </label>
-                {errors.joinWhatsappConfirmed && <div className="error-message">{errors.joinWhatsappConfirmed}</div>}
-              </div>
-            </div>
-
-            <div className="form-section">
-              <h2 className="form-section-title">&gt;&gt;&gt; Payment Section</h2>
-
-              <div className="form-group">
-                <label className="form-label required">Mode of Payment</label>
-                <div className="mcq-group">
-                  <label className="mcq-option">
-                    <input
-                      type="radio"
-                      name="paymentMode"
-                      value="cash"
-                      checked={formData.paymentMode === 'cash'}
-                      onChange={handleInputChange}
-                    />
-                    <span className="mcq-option-label">Cash</span>
-                  </label>
-                  <label className="mcq-option">
-                    <input
-                      type="radio"
-                      name="paymentMode"
-                      value="online"
-                      checked={formData.paymentMode === 'online'}
-                      onChange={handleInputChange}
-                    />
-                    <span className="mcq-option-label">Online</span>
-                  </label>
+                  {errors.fullName && <div className="error-message">{errors.fullName}</div>}
                 </div>
-                {errors.paymentMode && <div className="error-message">{errors.paymentMode}</div>}
-              </div>
 
-              <div className="form-group">
-                <label className="form-label required">Payment Date</label>
-                <input
-                  type="date"
-                  name="transactionDate"
-                  value={formData.transactionDate}
-                  onChange={handleInputChange}
-                  className="retro-input"
-                  max={new Date().toISOString().split('T')[0]}
-                />
-                {errors.transactionDate && <div className="error-message">{errors.transactionDate}</div>}
-              </div>
-
-              {formData.paymentMode === 'online' && (
-                <>
-                  <div className="form-group">
-                    <label className="form-label required">Transaction ID (For Online Payment)</label>
-                    <input
-                      type="text"
-                      name="transactionId"
-                      value={formData.transactionId}
-                      onChange={handleInputChange}
-                      className="retro-input"
-                      placeholder="Transaction ID"
-                    />
-                    {errors.transactionId && <div className="error-message">{errors.transactionId}</div>}
-                  </div>
-
-                  <div className="form-group">
-                    <label className="form-label required">Upload Payment Screenshot</label>
-                    <div className="file-upload-wrapper">
-                      <div className="file-upload">
+                <div className="form-group">
+                  <label className="form-label required">Year</label>
+                  <div className="mcq-group">
+                    {YEAR_OPTIONS.map((option) => (
+                      <label className="mcq-option" key={option}>
                         <input
-                          type="file"
-                          name="paymentScreenshot"
-                          id="paymentScreenshot"
-                          className="file-upload-input"
-                          accept="image/jpeg,image/jpg,image/png,.jpg,.jpeg,.png"
+                          type="radio"
+                          name="year"
+                          value={option}
+                          checked={formData.year === option}
                           onChange={handleInputChange}
                         />
-                        <label htmlFor="paymentScreenshot" className="file-upload-label">
-                          <div className="file-upload-icon"></div>
-                          <div className="file-upload-text">
-                            <span className="highlight">Click to upload</span>
-                            <br />
-                            PNG, JPG, JPEG
-                          </div>
-                        </label>
-                      </div>
-                      {formData.paymentScreenshot && <div className="file-name">{formData.paymentScreenshot.name}</div>}
-                    </div>
-                    {errors.paymentScreenshot && <div className="error-message">{errors.paymentScreenshot}</div>}
+                        <span className="mcq-option-label">{option}</span>
+                      </label>
+                    ))}
                   </div>
-                </>
-              )}
+                  {errors.year && <div className="error-message">{errors.year}</div>}
+                </div>
 
-              {formData.paymentMode === 'cash' && (
                 <div className="form-group">
-                  <label className="form-label required">Upload Cash Receipt / Slip</label>
+                  <label className="form-label required">Stream / Branch</label>
+                  <div className="mcq-group">
+                    {STREAM_OPTIONS.map((option) => (
+                      <label className="mcq-option" key={option}>
+                        <input
+                          type="radio"
+                          name="streamBranch"
+                          value={option}
+                          checked={formData.streamBranch === option}
+                          onChange={handleInputChange}
+                        />
+                        <span className="mcq-option-label">{option}</span>
+                      </label>
+                    ))}
+                  </div>
+                  {errors.streamBranch && <div className="error-message">{errors.streamBranch}</div>}
+                </div>
+
+                <div className="form-group">
+                  <label className="form-label required">College Name</label>
+                  <div className="mcq-group">
+                    {COLLEGE_OPTIONS.map((option) => (
+                      <label className="mcq-option" key={option}>
+                        <input
+                          type="radio"
+                          name="collegeName"
+                          value={option}
+                          checked={formData.collegeName === option}
+                          onChange={handleInputChange}
+                        />
+                        <span className="mcq-option-label">{option}</span>
+                      </label>
+                    ))}
+                  </div>
+                  {errors.collegeName && <div className="error-message">{errors.collegeName}</div>}
+                </div>
+
+                {formData.collegeName === 'Others' && (
+                  <div className="form-group">
+                    <label className="form-label required">Specify College Name</label>
+                    <input
+                      type="text"
+                      name="collegeOther"
+                      value={formData.collegeOther}
+                      onChange={handleInputChange}
+                      className="retro-input"
+                      placeholder="Enter your college name"
+                    />
+                    {errors.collegeOther && <div className="error-message">{errors.collegeOther}</div>}
+                  </div>
+                )}
+
+                <div className="form-group">
+                  <label className="form-label required">College ID / Library Card Upload</label>
                   <div className="file-upload-wrapper">
                     <div className="file-upload">
                       <input
                         type="file"
-                        name="cashReceipt"
-                        id="cashReceipt"
+                        name="collegeIdProof"
+                        id="collegeIdProof"
                         className="file-upload-input"
                         accept="image/jpeg,image/jpg,image/png,.jpg,.jpeg,.png"
                         onChange={handleInputChange}
                       />
-                      <label htmlFor="cashReceipt" className="file-upload-label">
-                        <div className="file-upload-icon"></div>
+                      <label htmlFor="collegeIdProof" className="file-upload-label">
+                        <div className="file-upload-icon">📁</div>
                         <div className="file-upload-text">
                           <span className="highlight">Click to upload</span>
                           <br />
@@ -421,52 +290,275 @@ const KhetRegistration = () => {
                         </div>
                       </label>
                     </div>
-                    {formData.cashReceipt && <div className="file-name">{formData.cashReceipt.name}</div>}
+                    {formData.collegeIdProof && <div className="file-name">✓ {formData.collegeIdProof.name}</div>}
                   </div>
-                  {errors.cashReceipt && <div className="error-message">{errors.cashReceipt}</div>}
+                  {errors.collegeIdProof && <div className="error-message">{errors.collegeIdProof}</div>}
                 </div>
-              )}
-            </div>
 
-            <div className="form-section">
-              <h2 className="form-section-title">&gt;&gt;&gt; Mandatory Declaration</h2>
-              <div className="form-group">
-                <label className="checkbox-group">
+                <div className="form-group">
+                  <label className="form-label required">Contact Number (Preferably WhatsApp Number)</label>
                   <input
-                    type="checkbox"
-                    name="declarationConfirmed"
-                    checked={formData.declarationConfirmed}
+                    type="text"
+                    name="contactNumber"
+                    value={formData.contactNumber}
                     onChange={handleInputChange}
+                    className="retro-input"
+                    placeholder="Contact Number"
                   />
-                  <span className="checkbox-custom"></span>
-                  <span className="checkbox-label">
-                    I have checked all the details carefully and have joined the WhatsApp group
-                  </span>
-                </label>
-                {errors.declarationConfirmed && <div className="error-message">{errors.declarationConfirmed}</div>}
-              </div>
-            </div>
+                  {errors.contactNumber && <div className="error-message">{errors.contactNumber}</div>}
+                </div>
 
-            <div className="submit-button-wrapper">
-              <button type="submit" className="retro-button" disabled={isSubmitting}>
-                {isSubmitting ? 'Submitting...' : 'Submit Registration'}
-              </button>
-              <button
-                type="button"
-                className="retro-button secondary"
-                onClick={() => history.goBack()}
-                disabled={isSubmitting}
-              >
-                Cancel
-              </button>
-            </div>
-
-            {submitSuccess && (
-              <div className="success-message" style={{ marginTop: '20px' }}>
-                ✓ Registration Successful! Redirecting to events page...
+                <div className="form-group">
+                  <label className="form-label required">Email Address</label>
+                  <input
+                    type="text"
+                    name="emailAddress"
+                    value={formData.emailAddress}
+                    onChange={handleInputChange}
+                    className="retro-input"
+                    placeholder="Email Address"
+                  />
+                  {errors.emailAddress && <div className="error-message">{errors.emailAddress}</div>}
+                </div>
               </div>
-            )}
-          </form>
+            </Step>
+
+            {/* Step 2: Payment Section */}
+            <Step>
+              <div className="form-section">
+                <h2 className="form-section-title">&gt;&gt;&gt; Payment Section</h2>
+
+                <div style={{
+                  backgroundColor: 'rgba(255, 192, 16, 0.1)',
+                  border: '2px solid #ffc010',
+                  padding: '20px',
+                  borderRadius: '5px',
+                  marginBottom: '25px'
+                }}>
+                  <h4 style={{ color: '#ffc010', marginTop: 0, marginBottom: '15px', fontSize: '14px' }}>
+                    💰 Registration Fee
+                  </h4>
+                  <p style={{ color: '#fff', lineHeight: '1.8', margin: 0, fontSize: '16px' }}>
+                    <strong style={{ color: '#ffc010' }}>Fee:</strong> ₹50 per participant
+                    <br />
+                    <span style={{ color: '#ff6b6b', fontSize: '14px' }}>⚠️ This event is only for BPPIMT students</span>
+                  </p>
+                </div>
+
+                <div className="form-group">
+                  <label className="form-label required">Mode of Payment</label>
+                  <div className="mcq-group">
+                    <label className="mcq-option">
+                      <input
+                        type="radio"
+                        name="paymentMode"
+                        value="cash"
+                        checked={formData.paymentMode === 'cash'}
+                        onChange={handleInputChange}
+                      />
+                      <span className="mcq-option-label">Cash</span>
+                    </label>
+                    <label className="mcq-option">
+                      <input
+                        type="radio"
+                        name="paymentMode"
+                        value="online"
+                        checked={formData.paymentMode === 'online'}
+                        onChange={handleInputChange}
+                      />
+                      <span className="mcq-option-label">Online</span>
+                    </label>
+                  </div>
+                  {errors.paymentMode && <div className="error-message">{errors.paymentMode}</div>}
+                </div>
+
+                {formData.paymentMode === 'online' && (
+                  <>
+                    {/* QR Code and UPI ID Section */}
+                    <div className="payment-qr-section">
+                      <div className="payment-qr-title">SCAN QR CODE TO PAY</div>
+                      <div className="payment-qr-container">
+                        <div className="qr-code-wrapper">
+                          <img 
+                            src={qrCodeImage} 
+                            alt="Payment QR Code" 
+                            className="qr-code-image"
+                          />
+                        </div>
+                        <div className="upi-id-container">
+                          <div className="upi-id-label">UPI ID</div>
+                          <div className="upi-id-value">bppoddar@iob</div>
+                        </div>
+                        <div className="payment-instruction-note">
+                          Scan the QR code or use the UPI ID to make payment
+                        </div>
+                      </div>
+                    </div>
+
+                    <div className="form-group">
+                      <label className="form-label required">Payment Date</label>
+                      <input
+                        type="date"
+                        name="transactionDate"
+                        value={formData.transactionDate}
+                        onChange={handleInputChange}
+                        className="retro-input"
+                        max={new Date().toISOString().split('T')[0]}
+                      />
+                      {errors.transactionDate && <div className="error-message">{errors.transactionDate}</div>}
+                    </div>
+
+                    <div className="form-group">
+                      <label className="form-label required">Transaction ID (For Online Payment)</label>
+                      <input
+                        type="text"
+                        name="transactionId"
+                        value={formData.transactionId}
+                        onChange={handleInputChange}
+                        className="retro-input"
+                        placeholder="Transaction ID"
+                      />
+                      {errors.transactionId && <div className="error-message">{errors.transactionId}</div>}
+                    </div>
+
+                    <div className="form-group">
+                      <label className="form-label required">Upload Payment Screenshot</label>
+                      <div className="file-upload-wrapper">
+                        <div className="file-upload">
+                          <input
+                            type="file"
+                            name="paymentScreenshot"
+                            id="paymentScreenshot"
+                            className="file-upload-input"
+                            accept="image/jpeg,image/jpg,image/png,.jpg,.jpeg,.png"
+                            onChange={handleInputChange}
+                          />
+                          <label htmlFor="paymentScreenshot" className="file-upload-label">
+                            <div className="file-upload-icon">💳</div>
+                            <div className="file-upload-text">
+                              <span className="highlight">Click to upload</span>
+                              <br />
+                              PNG, JPG, JPEG
+                            </div>
+                          </label>
+                        </div>
+                        {formData.paymentScreenshot && <div className="file-name">✓ {formData.paymentScreenshot.name}</div>}
+                      </div>
+                      {errors.paymentScreenshot && <div className="error-message">{errors.paymentScreenshot}</div>}
+                    </div>
+                  </>
+                )}
+
+                {formData.paymentMode === 'cash' && (
+                  <div className="form-group">
+                    <div style={{
+                      backgroundColor: 'rgba(255, 192, 16, 0.1)',
+                      border: '2px solid #ffc010',
+                      padding: '20px',
+                      borderRadius: '5px',
+                      marginTop: '20px'
+                    }}>
+                      <h4 style={{ color: '#ffc010', marginTop: 0, marginBottom: '15px' }}>
+                        📌 Important: Cash Payment Instructions
+                      </h4>
+                      <p style={{ color: '#fff', lineHeight: '1.8', margin: 0 }}>
+                        You must pay the registration fee in cash within <strong style={{ color: '#ffc010' }}>7 days of registration</strong> at the <strong style={{ color: '#ffc010' }}>Alumni Room</strong> of the college. 
+                        <br /><br />
+                        Please collect the paper receipt during payment and <strong style={{ color: '#ffc010' }}>keep it safe for event day verification</strong>.
+                        <br /><br />
+                        <span style={{ color: '#ff6b6b' }}>⚠️ No receipt upload is required during registration.</span>
+                      </p>
+                    </div>
+                  </div>
+                )}
+              </div>
+            </Step>
+
+            {/* Step 3: Rules & Agreement */}
+            <Step>
+              <div className="form-section">
+                <h2 className="form-section-title">&gt;&gt;&gt; Rules & Regulations</h2>
+                
+                <div className="rules-container" style={{ 
+                  background: '#1a1a1a', 
+                  padding: '30px', 
+                  borderRadius: '8px',
+                  marginBottom: '30px',
+                  border: '2px solid #333'
+                }}>
+                  <h3 style={{ color: '#ffc010', marginBottom: '20px' }}>Khet Event Rules</h3>
+                  <ul style={{ color: '#ccc', lineHeight: '1.8', paddingLeft: '20px' }}>
+                    <li>All participants must carry a valid college ID card</li>
+                    <li>This is an individual event - only 1 participant allowed</li>
+                    <li>Registration is mandatory for all participants</li>
+                    <li>Participants must follow the event schedule strictly</li>
+                    <li>Any form of malpractice will lead to disqualification</li>
+                    <li>Organizers' decision will be final and binding</li>
+                    <li>Entry fee is non-refundable under any circumstances</li>
+                  </ul>
+                </div>
+
+                <div className="form-group">
+                  <label className="form-label required">Join Official WhatsApp Group</label>
+                  <p style={{ margin: '10px 0' }}>
+                    <a
+                      href="https://chat.whatsapp.com/LeNZPNivID87ydepOLpAvY?mode=gi_t"
+                      target="_blank"
+                      rel="noopener noreferrer"
+                      style={{ color: '#ffc010' }}
+                    >
+                      https://chat.whatsapp.com/LeNZPNivID87ydepOLpAvY?mode=gi_t
+                    </a>
+                  </p>
+                  <label className="checkbox-group" style={{ marginTop: '12px' }}>
+                    <input
+                      type="checkbox"
+                      name="joinWhatsappConfirmed"
+                      checked={formData.joinWhatsappConfirmed}
+                      onChange={handleInputChange}
+                    />
+                    <span className="checkbox-custom"></span>
+                    <span className="checkbox-label">I have joined the official WhatsApp group</span>
+                  </label>
+                  {errors.joinWhatsappConfirmed && <div className="error-message">{errors.joinWhatsappConfirmed}</div>}
+                </div>
+
+                <div className="form-group">
+                  <label className="checkbox-group">
+                    <input
+                      type="checkbox"
+                      name="declarationConfirmed"
+                      checked={formData.declarationConfirmed}
+                      onChange={handleInputChange}
+                    />
+                    <span className="checkbox-custom"></span>
+                    <span className="checkbox-label">
+                      I have checked all the details carefully and have joined the WhatsApp group
+                    </span>
+                  </label>
+                  {errors.declarationConfirmed && <div className="error-message">{errors.declarationConfirmed}</div>}
+                </div>
+              </div>
+            </Step>
+          </Stepper>
+
+          {submitSuccess && (
+            <div className="success-message" style={{ marginTop: '20px', textAlign: 'center', padding: '20px', background: '#1a1a1a', borderRadius: '8px' }}>
+              <h3 style={{ color: '#ffc010' }}>✓ Registration Successful!</h3>
+              <p style={{ color: '#999' }}>Redirecting to events page...</p>
+            </div>
+          )}
+
+          <div style={{ marginTop: '20px', textAlign: 'center' }}>
+            <button 
+              type="button" 
+              className="retro-button secondary"
+              onClick={() => history.goBack()}
+              disabled={isSubmitting}
+            >
+              Cancel Registration
+            </button>
+          </div>
         </div>
       </div>
     </div>
