@@ -5,6 +5,7 @@ import {
   useCallback,
   useLayoutEffect,
 } from "react";
+import aboutBg from "../../../assets/img/aboutandform.png";
 import gsap from "gsap";
 import { ScrollTrigger } from "gsap/ScrollTrigger";
 
@@ -963,8 +964,10 @@ export default function About() {
         position: "relative",
         height: SECTION_H,
         minHeight: isMobile ? "auto" : "auto",
-        background:
-          "linear-gradient(180deg,#0A0A0A 0%,#0A0E2E 40%,#0d0520 70%,#0A0A0A 100%)",
+        backgroundImage: `linear-gradient(180deg, rgba(10,10,10,0.65) 0%, rgba(10,14,46,0.62) 40%, rgba(13,5,32,0.62) 70%, rgba(10,10,10,0.70) 100%), url(${aboutBg})`,
+        backgroundSize: "cover",
+        backgroundPosition: "center center",
+        backgroundAttachment: isMobile ? "scroll" : "fixed",
         paddingBottom: isMobile ? "60px" : "0",
         overflow: isMobile ? "hidden" : "clip",
       }}
